Moves Text component to shared
