+++ src/components/shared/Text/Text.js
@@ -1,22 +1,25 @@
-import React from "react";
+import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import "./Text.scss";
+import './Text.scss';
 
 const Text = (props) => {
-  const textClasses = cx("Text", {
-    "Text--large": props.fontSize === "large",
-    "Text--xl": props.fontSize === "xl"
+  const textClasses = cx('Text', {
+    'Text--large': props.fontSize === 'large',
+    'Text--xl': props.fontSize === 'xl'
   });
   return (
     <p className={textClasses}>{props.children}</p>
   );
 };
+
 Text.propTypes = {
   children: PropTypes.node.isRequired,
   fontSize: PropTypes.oneOf(['body', 'large', 'xl'])
 };
+
 Text.defaultProps = {
   fontSize: 'body'
 };
+
 export default Text;
